+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## master
 * Fix default contact form setting for Gutenberg contact form block
+* Store booleans as-is - fixes certain Gutenberg toggles on re-edit
 
 ## 2.3.3 (2019-03-26)
 * Fix new online booking in IE